% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lehdr.R
\name{aggregate_lodes_df}
\alias{aggregate_lodes_df}
\title{Aggregate data to a certain level dictated by inputs. Internal function.}
\usage{
aggregate_lodes_df(lehdr_df, geoid_to, aggname)
}
\arguments{
\item{lehdr_df}{Data frame (tibble) to be aggregated}

\item{geoid_to}{The number of characters to do the aggregation on from the geoid}

\item{aggname}{What the level is called, like "tract" etc}
}
\description{
Helper function for lehdr which aggregates block geographies based on the block id.
}
