% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lehdr.R
\name{grab_wac}
\alias{grab_wac}
\title{Download and load LEHD WAC data into a data frame (tibble)}
\usage{
grab_wac(state, year = 2014, segment = "S000", jobtype = "JT00",
  tract = FALSE)
}
\arguments{
\item{state}{US state abbrevation in lower case.}

\item{year}{year of the wac data.}

\item{segment}{...}

\item{jobtype}{...}

\item{tract}{logical whether to aggregate w_geocode to w_tract_id.}
}
\value{
A data frame containing block or tract level WAC.
}
\description{
\code{grab_wac} downloads and loads LEHD WAC data into a data frame (tibble)
}
