% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grab_lodes.R
\name{grab_lodes}
\alias{grab_lodes}
\title{Download and load LODES data into a data frame (tibble)}
\usage{
grab_lodes(state, year, lodes_type = c("od", "rac", "wac"), job_type, segment,
  part, tract = FALSE, download_dir = getwd())
}
\arguments{
\item{state}{US state abbreviation in lower case}

\item{year}{year of the lodes data}

\item{lodes_type}{table type: "origin-destination", "residential association", 
or "workplace association". OD files give a home and destination census block
for workers. Residential association files give job totals of worker home census 
blocks and workplace association files give job totals of worker job census blocks}

\item{job_type}{Jobtype: "JT00" for all jobs, "JT01" for Primary Jobs, "JT02" for All Private Jobs,
"JT03" for Private Primary Jobs, "JT04" for All Federal Jobs, "JT05" for Federal Primary Jobs}

\item{segment}{Segment of the workforce: "S000" total number of jobs for workers,
"SA01" number of jobs forworker aged 29 or younger, "SA02" number of jobs for workers 
aged 30-54,"SA03" number of jobs for workers 55 and older, "SE01" number of jobs 
with earnings $1,250/month or less, "SE02" number of jobs with earnings $1,251 to 
$3,333/month, "SE03" number of jobs with earnings greater than $3,333/month, 
"SI01" number of jobs in Goods Producing industry sectors, "SI02" number of jobs in
Trade, Transportation, and Utilities industry sectors, "SI03" number of jobs in All Other Services 
industry sectors}

\item{part}{Part of the state file, can have values of "main" or "aux" in OD files. "Main"
includes workers with their workplace and residence in the state. "Aux" includes workers with 
residences out of state and workplace in the state of interest}

\item{tract}{logical whether to aggregate w_geocode to w_tract_id.}

\item{download_dir}{Directory where lodes table will be downloaded}
}
\value{
a dataframe (tibble) of block or tract level LODES files
}
\description{
Download LODES OD, RAC, and WAC tables
}
