% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lehdr.R
\name{grab_rac}
\alias{grab_rac}
\title{Download and load LEHD RAC data into a data frame (tibble)}
\usage{
grab_rac(state, year = 2014, segment = "S000", jobtype = "JT00",
  tract = FALSE)
}
\arguments{
\item{state}{US state abbrevation in lower case.}

\item{year}{year of the wac data.}

\item{segment}{TODO}

\item{jobtype}{TODO}

\item{tract}{logical whether to aggregate h_geocode to h_tract_id.}
}
\value{
A data frame containing block or tract level WAC.
}
\description{
TODO
}
