% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lehdr.R
\name{grab_rac}
\alias{grab_rac}
\title{Download and load LEHD RAC data into a data frame (tibble)}
\usage{
grab_rac(state, year = 2014, segment = "S000", jobtype = "JT00",
  tract = FALSE)
}
\arguments{
\item{state}{US state abbrevation in lower case.}

\item{year}{year of the wac data.}

\item{segment}{Segment of the workforce: "S000" total number of jobs for workers,
"SA01" number of jobs forworker aged 29 or younger, "SA02" number of jobs for workers aged 30-54,
"SA03" number of jobs for workers 55 and older, "SE01" number of jobs with earnings $1,250/month or less,
"SE02" number of jobs with earnings $1,251 to $3,333/month, "SE03" number of jobs with earnings greater
than $3,333/month, "SI01" number of jobs in Goods Producing industry sectors, "SI02" number of jobs in
Trade, Transportation, and Utilities industry sectors, "SI03" number of jobs in All Other Services 
industry sectors}

\item{jobtype}{Jobtype: "JT00" for all jobs, "JT01" for Primary Jobs, "JT02" for All Private Jobs,
"JT03" for Private Primary Jobs, "JT04" for All Federal Jobs, "JT05" for Federal Primary Jobs}

\item{tract}{logical whether to aggregate h_geocode to h_tract_id.}
}
\value{
A data frame containing block or tract level WAC.
}
\description{
TODO
}
